don't use Long#intValue() in order to compare a Long instance with a long value
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@3703 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/explicit/ExplicitElement.java
@@ -80,7 +80,7 @@ public class ExplicitElement implements AbstractQueryableElement, FormulaReporti
   public void assignConstant(String nameOfVar, long value, int pThreshold){
 
     if(constantsMap.containsKey(nameOfVar) &&
-        constantsMap.get(nameOfVar).intValue() == value){
+        constantsMap.get(nameOfVar).longValue() == value){
       return;
     }
